+++ apps/package-manager/packages/generic/src/packageManager.ts
@@ -114,6 +114,7 @@ export class PackageManagerHandler {
 	onSettingsChanged(): void {
 		this.settings = {
 			delayRemoval: this.coreHandler.delayRemoval,
+			delayRemovalPackageInfo: this.coreHandler.delayRemovalPackageInfo,
 			useTemporaryFilePath: this.coreHandler.useTemporaryFilePath,
 		}
 		this.triggerUpdatedExpectedPackages()
@@ -268,7 +269,7 @@ export class PackageManagerHandler {
 			this.settings
 		)
 		this.logger.debug(`Has ${Object.keys(expectations).length} expectations`)
-		// this.logger.debug(JSON.stringify(expectations, null, 2))
+		console.log(JSON.stringify(expectations, null, 2))
 		this.dataSnapshot.expectations = expectations
 
 		const packageContainerExpectations = generatePackageContainerExpectations(
@@ -915,6 +916,7 @@ export interface ActiveRundown {
 }
 export interface PackageManagerSettings {
 	delayRemoval: number
+	delayRemovalPackageInfo?: number
 	useTemporaryFilePath: boolean
 }